fakermaker: refactor core functions into package
